Lib

Documentation

data PuzzleState

PuzzleState contains the current moves (fn), distance to goal (gn), current position of blank tile (zeroPos), and the current board state (state)

Constructors

PuzzleState

Fields

fn :: Int

gn :: Int

zeroPos :: Int

state :: Array U DIM1 Int

generateArrays :: (Num a, Enum a) => Int -> a -> [[a]]

generateArrays returns k number of shuffled matrix of size n for the input of 15-puzzle problem

formatArray :: [Int] -> Int -> String

formatArray takes the array (a) and the size of the puzzle (n) and return it as a string, according to the input text format of this program

formatArrays :: [[Int]] -> String

formatArrays takes the arrays (a:as) and return it as a string according to the input text format of this program

writeArrays :: [[Int]] -> FilePath -> IO ()

writeArrays takes the arrays and write it into the filename according to the input text format of this program

manhattanDist :: Source r Int => Array r DIM1 Int -> Int -> Int -> Int

manhattanDist calculates the total distance of the current state (cur) to the goal board with size (n), performing recurrsion using (idx)

hammingDist :: Source r Int => Array r DIM1 Int -> Int -> Int -> Int

hammingDist calculates the number of wrong tiles of the current state (cur) to the goal board with size (n), performing recursion using (idx)

readInt :: Handle -> IO Int

readInt parse the input handle and return an Integer from its first line

printList :: Show a => [a] -> IO ()

printList print a given list (l) into IO

getStateVector :: Handle -> Int -> Int -> IO [[Int]]

getStateVector parse the input handle and return lists of list of integer, which is the initial game board

getZeroPos :: Source r Int => Array r DIM1 Int -> Int -> Int

getZeroPos returns the idx within the given array (arr) where the blank tile is located. If fail, return -1

swapTwo :: Source r Int => Int -> Int -> Array r DIM1 Int -> Array D DIM1 Int

swapTwo perform swap between two elements in the array (arr), given two indexes, f and s in the array

getUpNeighbor :: PuzzleState -> Int -> Maybe PuzzleState

getUpNeighbor return the subsequent PuzzleState by swapping the blank tile with the tile above it. If its impossible, return Nothing

getDownNeighbor :: PuzzleState -> Int -> Maybe PuzzleState

getDownNeighbor return the subsequent PuzzleState by swapping the blank tile with the tile below it. If its impossible, return Nothing

getLeftNeighbor :: PuzzleState -> Int -> Maybe PuzzleState

getLeftNeighbor return the subsequent PuzzleState by swapping the blank tile with the tile left to it. If its impossible, return Nothing

getRightNeighbor :: PuzzleState -> Int -> Maybe PuzzleState

getRightNeighbor return the subsequent PuzzleState by swapping the blank tile with the tile right to it. If its impossible, return Nothing

getAllNeighbor :: PuzzleState -> Int -> [PuzzleState]

getAllNeighbor return all of the neighboring state of the current PuzzleState

getValidNeighbor :: [PuzzleState] -> HashMap String Int -> [PuzzleState]

getValidNeighbor filters all neighbor puzzles that improves (fn) or have not been discovered previously (not in mp)

filterInMap :: HashMap [Char] Int -> PuzzleState -> Bool

filterInMap returns True if the puzzle (puzzle) is not in the HashMap (mp) or if the puzzle can now be reached in less steps (fn)

addMap :: Foldable t => t PuzzleState -> HashMap [Char] Int -> HashMap [Char] Int

addMap add all of the puzzle states (ps) into the given HashMap (mp)

addPSQ :: Foldable t => t PuzzleState -> PSQ PuzzleState Int -> PSQ PuzzleState Int

addPSQ adds all of the given puzzle states (ps) into the PriorityQueue (psq)

getHashKey :: Array U DIM1 Int -> [Char]

getHashKey turns the hash result from the given array (li) and return string as the hash key. hash [0, 3, 1, 2] -> "00030102"

hash :: Integral a => Array U DIM1 Int -> Int -> a

hash perform simple hash function on the given array (l), using recursive function on idx. hash [0, 3, 1, 2] -> "00030102"

numinv :: Array U DIM1 Int -> Int

numinv check the number of inversions in the board (arr)

solvability :: Array U DIM1 Int -> Int -> Int -> Bool

solvability checks whether the given board (arr) with the current zero position (zeropos) is solvable 8-puzzle problem

solveKpuzzle :: Handle -> Int -> IO ()

solveKpuzzle perform solving on multiple 8-puzzles using A* algorithm

solve :: PSQ PuzzleState Int -> Array U DIM1 Int -> Int -> HashMap String Int -> IO Int

solve perform sequential solving on 8-puzzle using A* algorithm